removed java static reference from translate by introducing idea of post processing
