+++ javascript/gast_to_code/js_gast_to_code_converter.py
@@ -20,6 +20,9 @@ class JsGastToCodeConverter():
     def code_to_gast(self, code_input):
         return js_main.js_to_gast(code_input)
 
+    def gast_to_code_post_processing(self, code_output):
+        return code_output  # since there is no post-processing
+
     def handle_bool(self, gast):
         if gast["value"] == 1:
             return "true"
